Configure live migration in Nova
For now this is the only working solution that doesn't require running scp
between computes (and configuring ssh keys) and should work without patching
Nova. But hostname has to be used in live_migration_uri.
More correct setting would be:
live_migration_inbound_addr = "{{ network_topology["private"]["address"] }}"
and no live_migration_uri (it's not required in this case)
But this won't work in master untill https://review.openstack.org/#/c/398956/
is merged and in Netwton until fixes are backported.
Also this patch adds dosfstools package to nova-compute container which is
needed for Nova to create config drive for an instance with mkfs.vfat command.

Change-Id: Ibeca3aca413c5ceace97b836bc56320fb9947220